requirements/base.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.3
drf-spectacular>=0.27
psycopg[binary]>=3.2
redis>=5.2
celery>=5.4
django-environ>=0.11
django-cors-headers>=4.4
sentry-sdk[django]>=2.17
gunicorn>=23.0; platform_system != "Windows"
uvicorn[standard]>=0.32
boto3>=1.35,<2
cryptography>=43,<46
